sizet draft

+++ src/cryptonote_core/cryptonote_basic_impl.cpp
@@ -63,17 +63,17 @@ namespace cryptonote {
   /* Cryptonote helper functions                                          */
   /************************************************************************/
   //-----------------------------------------------------------------------------------------------
-  size_t get_max_block_size()
+  uint64_t get_max_block_size()
   {
     return CRYPTONOTE_MAX_BLOCK_SIZE;
   }
   //-----------------------------------------------------------------------------------------------
-  size_t get_max_tx_size()
+  uint64_t get_max_tx_size()
   {
     return CRYPTONOTE_MAX_TX_SIZE;
   }
   //-----------------------------------------------------------------------------------------------
-  bool get_block_reward(size_t median_size, size_t current_block_size, uint64_t already_generated_coins, uint64_t &reward, uint8_t version) {
+  bool get_block_reward(uint64_t median_size, uint64_t current_block_size, uint64_t already_generated_coins, uint64_t &reward, uint8_t version) {
     static_assert(DIFFICULTY_TARGET_V2%60==0&&DIFFICULTY_TARGET_V1%60==0,"difficulty targets must be a multiple of 60");
     const int target = version < 2 ? DIFFICULTY_TARGET_V1 : DIFFICULTY_TARGET_V2;
     const int target_minutes = target / 60;
@@ -106,8 +106,6 @@ namespace cryptonote {
     assert(current_block_size < std::numeric_limits<uint32_t>::max());
 
     uint64_t product_hi;
-    // BUGFIX: 32-bit saturation bug (e.g. ARM7), the result was being
-    // treated as 32-bit by default.
     uint64_t multiplicand = 2 * median_size - current_block_size;
     multiplicand *= current_block_size;
     uint64_t product_lo = mul128(base_reward, multiplicand, &product_hi);
