+++ tests/core_tests/chaingen.cpp
@@ -50,7 +50,7 @@ using namespace epee;
 using namespace cryptonote;
 
 
-void test_generator::get_block_chain(std::vector<block_info>& blockchain, const crypto::hash& head, size_t n) const
+void test_generator::get_block_chain(std::vector<block_info>& blockchain, const crypto::hash& head, uint64_t n) const
 {
   crypto::hash curr = head;
   while (null_hash != curr && blockchain.size() < n)
@@ -68,7 +68,7 @@ void test_generator::get_block_chain(std::vector<block_info>& blockchain, const 
   std::reverse(blockchain.begin(), blockchain.end());
 }
 
-void test_generator::get_last_n_block_sizes(std::vector<size_t>& block_sizes, const crypto::hash& head, size_t n) const
+void test_generator::get_last_n_block_sizes(std::vector<uint64_t>& block_sizes, const crypto::hash& head, uint64_t n) const
 {
   std::vector<block_info> blockchain;
   get_block_chain(blockchain, head, n);
@@ -94,9 +94,9 @@ uint64_t test_generator::get_already_generated_coins(const cryptonote::block& bl
   return get_already_generated_coins(blk_hash);
 }
 
-void test_generator::add_block(const cryptonote::block& blk, size_t tsx_size, std::vector<size_t>& block_sizes, uint64_t already_generated_coins, uint8_t hf_version)
+void test_generator::add_block(const cryptonote::block& blk, uint64_t tsx_size, std::vector<uint64_t>& block_sizes, uint64_t already_generated_coins, uint8_t hf_version)
 {
-  const size_t block_size = tsx_size + get_object_blobsize(blk.miner_tx);
+  const uint64_t block_size = tsx_size + get_object_blobsize(blk.miner_tx);
   uint64_t block_reward;
   get_block_reward(misc_utils::median(block_sizes), block_size, already_generated_coins, block_reward, hf_version);
   m_blocks_info[get_block_hash(blk)] = block_info(blk.prev_id, already_generated_coins + block_reward, block_size);
@@ -104,7 +104,7 @@ void test_generator::add_block(const cryptonote::block& blk, size_t tsx_size, st
 
 bool test_generator::construct_block(cryptonote::block& blk, uint64_t height, const crypto::hash& prev_id,
                                      const cryptonote::account_base& miner_acc, uint64_t timestamp, uint64_t already_generated_coins,
-                                     std::vector<size_t>& block_sizes, const std::list<cryptonote::transaction>& tx_list)
+                                     std::vector<uint64_t>& block_sizes, const std::list<cryptonote::transaction>& tx_list)
 {
   blk.major_version = CURRENT_BLOCK_MAJOR_VERSION;
   blk.minor_version = CURRENT_BLOCK_MINOR_VERSION;
@@ -120,7 +120,7 @@ bool test_generator::construct_block(cryptonote::block& blk, uint64_t height, co
   }
 
   uint64_t total_fee = 0;
-  size_t txs_size = 0;
+  uint64_t txs_size = 0;
   BOOST_FOREACH(auto& tx, tx_list)
   {
     uint64_t fee = 0;
@@ -131,20 +131,20 @@ bool test_generator::construct_block(cryptonote::block& blk, uint64_t height, co
   }
 
   blk.miner_tx = AUTO_VAL_INIT(blk.miner_tx);
-  size_t target_block_size = txs_size + get_object_blobsize(blk.miner_tx);
+  uint64_t target_block_size = txs_size + get_object_blobsize(blk.miner_tx);
   while (true)
   {
     if (!construct_miner_tx(height, misc_utils::median(block_sizes), already_generated_coins, target_block_size, total_fee, miner_acc.get_keys().m_account_address, blk.miner_tx, blobdata(), 10))
       return false;
 
-    size_t actual_block_size = txs_size + get_object_blobsize(blk.miner_tx);
+    uint64_t actual_block_size = txs_size + get_object_blobsize(blk.miner_tx);
     if (target_block_size < actual_block_size)
     {
       target_block_size = actual_block_size;
     }
     else if (actual_block_size < target_block_size)
     {
-      size_t delta = target_block_size - actual_block_size;
+      uint64_t delta = target_block_size - actual_block_size;
       blk.miner_tx.extra.resize(blk.miner_tx.extra.size() + delta, 0);
       actual_block_size = txs_size + get_object_blobsize(blk.miner_tx);
       if (actual_block_size == target_block_size)
@@ -189,7 +189,7 @@ bool test_generator::construct_block(cryptonote::block& blk, uint64_t height, co
 
 bool test_generator::construct_block(cryptonote::block& blk, const cryptonote::account_base& miner_acc, uint64_t timestamp)
 {
-  std::vector<size_t> block_sizes;
+  std::vector<uint64_t> block_sizes;
   std::list<cryptonote::transaction> tx_list;
   return construct_block(blk, 0, null_hash, miner_acc, timestamp, 0, block_sizes, tx_list);
 }
@@ -203,7 +203,7 @@ bool test_generator::construct_block(cryptonote::block& blk, const cryptonote::b
   // Keep difficulty unchanged
   uint64_t timestamp = blk_prev.timestamp + DIFFICULTY_BLOCKS_ESTIMATE_TIMESPAN;
   uint64_t already_generated_coins = get_already_generated_coins(prev_id);
-  std::vector<size_t> block_sizes;
+  std::vector<uint64_t> block_sizes;
   get_last_n_block_sizes(block_sizes, prev_id, CRYPTONOTE_REWARD_BLOCKS_WINDOW);
 
   return construct_block(blk, height, prev_id, miner_acc, timestamp, already_generated_coins, block_sizes, tx_list);
@@ -215,7 +215,7 @@ bool test_generator::construct_block_manually(block& blk, const block& prev_bloc
                                               const crypto::hash& prev_id/* = crypto::hash()*/, const difficulty_type& diffic/* = 1*/,
                                               const transaction& miner_tx/* = transaction()*/,
                                               const std::vector<crypto::hash>& tx_hashes/* = std::vector<crypto::hash>()*/,
-                                              size_t txs_sizes/* = 0*/, size_t max_outs/* = 0*/, uint8_t hf_version/* = 1*/)
+                                              uint64_t txs_sizes/* = 0*/, uint64_t max_outs/* = 0*/, uint8_t hf_version/* = 1*/)
 {
   blk.major_version = actual_params & bf_major_ver ? major_ver : CURRENT_BLOCK_MAJOR_VERSION;
   blk.minor_version = actual_params & bf_minor_ver ? minor_ver : CURRENT_BLOCK_MINOR_VERSION;
@@ -225,9 +225,9 @@ bool test_generator::construct_block_manually(block& blk, const block& prev_bloc
   max_outs          = actual_params & bf_max_outs ? max_outs : 9999;
   hf_version        = actual_params & bf_hf_version ? hf_version : 1;
 
-  size_t height = get_block_height(prev_block) + 1;
+  uint64_t height = get_block_height(prev_block) + 1;
   uint64_t already_generated_coins = get_already_generated_coins(prev_block);
-  std::vector<size_t> block_sizes;
+  std::vector<uint64_t> block_sizes;
   get_last_n_block_sizes(block_sizes, get_block_hash(prev_block), CRYPTONOTE_REWARD_BLOCKS_WINDOW);
   if (actual_params & bf_miner_tx)
   {
@@ -235,7 +235,7 @@ bool test_generator::construct_block_manually(block& blk, const block& prev_bloc
   }
   else
   {
-    size_t current_block_size = txs_sizes + get_object_blobsize(blk.miner_tx);
+    uint64_t current_block_size = txs_sizes + get_object_blobsize(blk.miner_tx);
     // TODO: This will work, until size of constructed block is less then CRYPTONOTE_BLOCK_GRANTED_FULL_REWARD_ZONE
     if (!construct_miner_tx(height, misc_utils::median(block_sizes), already_generated_coins, current_block_size, 0, miner_acc.get_keys().m_account_address, blk.miner_tx, blobdata(), max_outs, hf_version))
       return false;
@@ -253,7 +253,7 @@ bool test_generator::construct_block_manually(block& blk, const block& prev_bloc
 
 bool test_generator::construct_block_manually_tx(cryptonote::block& blk, const cryptonote::block& prev_block,
                                                  const cryptonote::account_base& miner_acc,
-                                                 const std::vector<crypto::hash>& tx_hashes, size_t txs_size)
+                                                 const std::vector<crypto::hash>& tx_hashes, uint64_t txs_size)
 {
   return construct_block_manually(blk, prev_block, miner_acc, bf_tx_hashes, 0, 0, 0, crypto::hash(), 0, transaction(), tx_hashes, txs_size);
 }
@@ -262,15 +262,15 @@ bool test_generator::construct_block_manually_tx(cryptonote::block& blk, const c
 struct output_index {
     const cryptonote::txout_target_v out;
     uint64_t amount;
-    size_t blk_height; // block height
-    size_t tx_no; // index of transaction in block
-    size_t out_no; // index of out in transaction
-    size_t idx;
+    uint64_t blk_height; // block height
+    uint64_t tx_no; // index of transaction in block
+    uint64_t out_no; // index of out in transaction
+    uint64_t idx;
     bool spent;
     const cryptonote::block *p_blk;
     const cryptonote::transaction *p_tx;
 
-    output_index(const cryptonote::txout_target_v &_out, uint64_t _a, size_t _h, size_t tno, size_t ono, const cryptonote::block *_pb, const cryptonote::transaction *_pt)
+    output_index(const cryptonote::txout_target_v &_out, uint64_t _a, uint64_t _h, uint64_t tno, uint64_t ono, const cryptonote::block *_pb, const cryptonote::transaction *_pt)
         : out(_out), amount(_a), blk_height(_h), tx_no(tno), out_no(ono), idx(0), spent(false), p_blk(_pb), p_tx(_pt) { }
 
     output_index(const output_index &other)
@@ -297,9 +297,9 @@ struct output_index {
     }
 };
 
-typedef std::map<uint64_t, std::vector<size_t> > map_output_t;
+typedef std::map<uint64_t, std::vector<uint64_t> > map_output_t;
 typedef std::map<uint64_t, std::vector<output_index> > map_output_idx_t;
-typedef pair<uint64_t, size_t>  outloc_t;
+typedef pair<uint64_t, uint64_t>  outloc_t;
 
 namespace
 {
@@ -315,7 +315,7 @@ namespace
   }
 }
 
-bool init_output_indices(map_output_idx_t& outs, std::map<uint64_t, std::vector<size_t> >& outs_mine, const std::vector<cryptonote::block>& blockchain, const map_hash2tx_t& mtx, const cryptonote::account_base& from) {
+bool init_output_indices(map_output_idx_t& outs, std::map<uint64_t, std::vector<uint64_t> >& outs_mine, const std::vector<cryptonote::block>& blockchain, const map_hash2tx_t& mtx, const cryptonote::account_base& from) {
 
     BOOST_FOREACH (const block& blk, blockchain) {
         vector<const transaction*> vtx;
@@ -331,17 +331,17 @@ bool init_output_indices(map_output_idx_t& outs, std::map<uint64_t, std::vector<
 
         //vtx.insert(vtx.end(), blk.);
         // TODO: add all other txes
-        for (size_t i = 0; i < vtx.size(); i++) {
+        for (uint64_t i = 0; i < vtx.size(); i++) {
             const transaction &tx = *vtx[i];
 
-            for (size_t j = 0; j < tx.vout.size(); ++j) {
+            for (uint64_t j = 0; j < tx.vout.size(); ++j) {
                 const tx_out &out = tx.vout[j];
 
                 output_index oi(out.target, out.amount, boost::get<txin_gen>(*blk.miner_tx.vin.begin()).height, i, j, &blk, vtx[i]);
 
                 if (2 == out.target.which()) { // out_to_key
                     outs[out.amount].push_back(oi);
-                    size_t tx_global_idx = outs[out.amount].size() - 1;
+                    uint64_t tx_global_idx = outs[out.amount].size() - 1;
                     outs[out.amount][tx_global_idx].idx = tx_global_idx;
                     // Is out to me?
                     if (is_out_to_acc(from.get_keys(), boost::get<txout_to_key>(out.target), get_tx_pub_key_from_extra(tx), j)) {
@@ -358,7 +358,7 @@ bool init_output_indices(map_output_idx_t& outs, std::map<uint64_t, std::vector<
 bool init_spent_output_indices(map_output_idx_t& outs, map_output_t& outs_mine, const std::vector<cryptonote::block>& blockchain, const map_hash2tx_t& mtx, const cryptonote::account_base& from) {
 
     BOOST_FOREACH (const map_output_t::value_type &o, outs_mine) {
-        for (size_t i = 0; i < o.second.size(); ++i) {
+        for (uint64_t i = 0; i < o.second.size(); ++i) {
             output_index &oi = outs[o.first][o.second[i]];
 
             // construct key image for this output
@@ -384,14 +384,14 @@ bool init_spent_output_indices(map_output_idx_t& outs, map_output_t& outs_mine, 
     return true;
 }
 
-bool fill_output_entries(std::vector<output_index>& out_indices, size_t sender_out, size_t nmix, size_t& real_entry_idx, std::vector<tx_source_entry::output_entry>& output_entries)
+bool fill_output_entries(std::vector<output_index>& out_indices, uint64_t sender_out, uint64_t nmix, uint64_t& real_entry_idx, std::vector<tx_source_entry::output_entry>& output_entries)
 {
   if (out_indices.size() <= nmix)
     return false;
 
   bool sender_out_found = false;
-  size_t rest = nmix;
-  for (size_t i = 0; i < out_indices.size() && (0 < rest || !sender_out_found); ++i)
+  uint64_t rest = nmix;
+  for (uint64_t i = 0; i < out_indices.size() && (0 < rest || !sender_out_found); ++i)
   {
     const output_index& oi = out_indices[i];
     if (oi.spent)
@@ -421,7 +421,7 @@ bool fill_output_entries(std::vector<output_index>& out_indices, size_t sender_o
 }
 
 bool fill_tx_sources(std::vector<tx_source_entry>& sources, const std::vector<test_event_entry>& events,
-                     const block& blk_head, const cryptonote::account_base& from, uint64_t amount, size_t nmix)
+                     const block& blk_head, const cryptonote::account_base& from, uint64_t amount, uint64_t nmix)
 {
     map_output_idx_t outs;
     map_output_t outs_mine;
@@ -442,9 +442,9 @@ bool fill_tx_sources(std::vector<tx_source_entry>& sources, const std::vector<te
     bool sources_found = false;
     BOOST_REVERSE_FOREACH(const map_output_t::value_type o, outs_mine)
     {
-        for (size_t i = 0; i < o.second.size() && !sources_found; ++i)
+        for (uint64_t i = 0; i < o.second.size() && !sources_found; ++i)
         {
-            size_t sender_out = o.second[i];
+            uint64_t sender_out = o.second[i];
             const output_index& oi = outs[o.first][sender_out];
             if (oi.spent)
                 continue;
@@ -453,7 +453,7 @@ bool fill_tx_sources(std::vector<tx_source_entry>& sources, const std::vector<te
             ts.amount = oi.amount;
             ts.real_output_in_tx_index = oi.out_no;
             ts.real_out_tx_key = get_tx_pub_key_from_extra(*oi.p_tx); // incoming tx public key
-            size_t realOutput;
+            uint64_t realOutput;
             if (!fill_output_entries(outs[o.first], sender_out, nmix, realOutput, ts.outputs))
               continue;
 
@@ -481,7 +481,7 @@ bool fill_tx_destination(tx_destination_entry &de, const cryptonote::account_bas
 
 void fill_tx_sources_and_destinations(const std::vector<test_event_entry>& events, const block& blk_head,
                                       const cryptonote::account_base& from, const cryptonote::account_base& to,
-                                      uint64_t amount, uint64_t fee, size_t nmix, std::vector<tx_source_entry>& sources,
+                                      uint64_t amount, uint64_t fee, uint64_t nmix, std::vector<tx_source_entry>& sources,
                                       std::vector<tx_destination_entry>& destinations)
 {
   sources.clear();
@@ -512,7 +512,7 @@ void fill_nonce(cryptonote::block& blk, const difficulty_type& diffic, uint64_t 
     blk.timestamp++;
 }
 
-bool construct_miner_tx_manually(size_t height, uint64_t already_generated_coins,
+bool construct_miner_tx_manually(uint64_t height, uint64_t already_generated_coins,
                                  const account_public_address& miner_address, transaction& tx, uint64_t fee,
                                  keypair* p_txkey/* = 0*/)
 {
@@ -554,7 +554,7 @@ bool construct_miner_tx_manually(size_t height, uint64_t already_generated_coins
 
 bool construct_tx_to_key(const std::vector<test_event_entry>& events, cryptonote::transaction& tx, const block& blk_head,
                          const cryptonote::account_base& from, const cryptonote::account_base& to, uint64_t amount,
-                         uint64_t fee, size_t nmix)
+                         uint64_t fee, uint64_t nmix)
 {
   vector<tx_source_entry> sources;
   vector<tx_destination_entry> destinations;
@@ -575,7 +575,7 @@ transaction construct_tx_with_fee(std::vector<test_event_entry>& events, const b
 uint64_t get_balance(const cryptonote::account_base& addr, const std::vector<cryptonote::block>& blockchain, const map_hash2tx_t& mtx) {
     uint64_t res = 0;
     std::map<uint64_t, std::vector<output_index> > outs;
-    std::map<uint64_t, std::vector<size_t> > outs_mine;
+    std::map<uint64_t, std::vector<uint64_t> > outs_mine;
 
     map_hash2tx_t confirmed_txs;
     get_confirmed_txs(blockchain, mtx, confirmed_txs);
@@ -587,7 +587,7 @@ uint64_t get_balance(const cryptonote::account_base& addr, const std::vector<cry
         return false;
 
     BOOST_FOREACH (const map_output_t::value_type &o, outs_mine) {
-        for (size_t i = 0; i < o.second.size(); ++i) {
+        for (uint64_t i = 0; i < o.second.size(); ++i) {
             if (outs[o.first][o.second[i]].spent)
                 continue;
 
@@ -656,7 +656,7 @@ void test_chain_unit_base::register_callback(const std::string& cb_name, verify_
 {
   m_callbacks[cb_name] = cb;
 }
-bool test_chain_unit_base::verify(const std::string& cb_name, cryptonote::core& c, size_t ev_index, const std::vector<test_event_entry> &events)
+bool test_chain_unit_base::verify(const std::string& cb_name, cryptonote::core& c, uint64_t ev_index, const std::vector<test_event_entry> &events)
 {
   auto cb_it = m_callbacks.find(cb_name);
   if(cb_it == m_callbacks.end())
